refactor(group): extract shared link editor table
